+++ CHANGELOG.md
@@ -5,4 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+* Officers can sign in with their existing SPD username and password
+
 No changes yet